moved App out of index
